Split out common EditorTable base class
